Update the strings used for JQ
The query string for hitbtc has been logged as an issue with
Chronicle Labs. Other options are provided for "simple" queries for
single and few source tickers, e.g. if we only need bitfinex for
a single price-pair, do we always want to recall total ticker data?

The philosophy of the code is to request as much as is possible up
front and so we can revert as we gain confidence in the stack.