Include response body in collector data
This information is used for auditing purposes in Orcfax. The
response is decoded and added to the metadata returned by the
collector. If decoding fails the collector fails to minimize the
risk of data not making its way downstream. Basic tests are
provided to provide some assurance this code is doing what it is
supposed to. The response is no longer encoded as Base64 as it
introduced a double encoding issue when the message was received
by the validator.