Update README.md
A subtle change to what we are calling this suite. It reflects the wider usage of the term "oracle" we are using with in the Orcfax domain to mean all the component parts of a network that publishes real world data to a blockchain. The core functions are Collecting, Validating, and Publishing data. An oracle does all three. This library addresses the data collection component which is why I'd like to recommend this name change.