Avoid an error if a model doesn't exist
Currently a slice of models where one model doesn't exist will result in
an error preventing the processing of other models. This is unnecessary when
we can just log a model doesn't exist and move on. This commit should prevent
that small issue and further, probably provide safer access to information in
model/node slices.